[MOD] Añadidos lotes
