+++ quality_control_ext/models/stock.py
@@ -1,9 +1,6 @@
 # -*- encoding: utf-8 -*-
 ##############################################################################
 #
-#    Avanzosc - Advanced Open Source Consulting
-#    Copyright (C) 2011 - 2013 Avanzosc <http://www.avanzosc.com>
-#
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Affero General Public License as
 #    published by the Free Software Foundation, either version 3 of the
@@ -19,9 +16,9 @@
 #
 ##############################################################################
 
-from openerp.osv import orm
+from openerp.osv import orm, fields
 from openerp.tools.translate import _
-import time
+# import time
 
 
 class StockMove(orm.Model):
@@ -90,7 +87,7 @@ class StockMove(orm.Model):
                     origin = (str(move.production_id.name) + ' - ' +
                               move.product_id.name)
                 data = {
-                    'name': time.strftime('%Y%m%d %H%M%S'),
+                    # 'name': time.strftime('%Y%m%d %H%M%S'),
                     'test_template_id': template.id,
                     'origin': origin,
                     'enabled': True,
@@ -98,11 +95,10 @@ class StockMove(orm.Model):
                     'state': 'draft',
                     'product_qty': move.product_qty,
                     'stock_move_id': move.id,
+                    'lot_id': move.lot_ids and move.lot_ids.id,
+                    'picking_id': move.picking_id and move.picking_id.id,
+                    'production_id': move.production_id and move.production_id.id
                 }
-                if move.picking_id:
-                    data.update({'picking_id': move.picking_id.id})
-                if move.production_id:
-                    data.update({'production_id': move.production_id.id})
                 test_id = test_obj.create(cr, uid, data, context=context)
                 # Creo las líneas del test
                 if template.test_template_line_ids:
@@ -116,23 +112,16 @@ class StockMove(orm.Model):
                                 'test_id': test_id,
                                 'test_template_line_id': line.id,
                                 'min_value': line.min_value,
-                                'max_value': line.max_value
+                                'max_value': line.max_value,
+                                'uom_id': line.uom_id and line.uom_id.id,
+                                'proof_id': line.proof_id and line.proof_id.id,
+                                'method_id': (line.method_id and
+                                              line.method_id.id),
+                                'proof_type': line.type,
                             }
-                            if line.uom_id:
-                                data.update({'uom_id': line.uom_id.id})
-                            if line.proof_id:
-                                data.update({'proof_id': line.proof_id.id})
-                            if line.method_id:
-                                data.update({'method_id': line.method_id.id})
                             if line.type:
-                                data.update({'proof_type': line.type})
                                 if (line.type == 'qualitative' and
                                         line.valid_value_ids):
-                                    data.update({'actual_value_ql': (
-                                                len(line.valid_value_ids)
-                                                and line.valid_value_ids[0]
-                                                and line.valid_value_ids[0].id
-                                                or False)})
                                     my_data = []
                                     for val in line.valid_value_ids:
                                         my_data.append(val.id)
@@ -143,3 +132,112 @@ class StockMove(orm.Model):
                                                  context=context)
                         howmany -= 1
         return True
+
+
+class StockPicking(orm.Model):
+    _inherit = 'stock.picking'
+
+    def _count_tests(self, cr, uid, ids, name, arg, context=None):
+        res = {}
+        for test in self.browse(cr, uid, ids, context=context):
+            created = 0
+            done = 0
+            passed = 0
+            failed = 0
+            if test.qc_test_ids:
+                for line in test.qc_test_ids:
+                    created += 1
+                    if line.state not in ('draft', 'waiting'):
+                        done += 1
+                        if line.state == 'success':
+                            passed += 1
+                        elif line.state == 'failed':
+                            failed += 1
+            res[test.id] = {
+                'created_tests': created,
+                'done_tests': done,
+                'passed_tests': passed,
+                'failed_tests': failed,
+            }
+        return res
+
+    _columns = {
+        # Test asociados al albaran
+        'qc_test_ids': fields.one2many('qc.test', 'picking_id', 'Tests'),
+        # Numero de test creados
+        'created_tests': fields.function(_count_tests, string="Created Tests",
+                                         type="integer", multi="test_count"),
+        # Numero de test realizados
+        'done_tests': fields.function(_count_tests, string="Realized Tests",
+                                      type="integer", multi="test_count"),
+        # Numero de test OK
+        'passed_tests': fields.function(_count_tests, string="Tests OK",
+                                        type="integer", multi="test_count"),
+        # Numero de test fallados
+        'failed_tests': fields.function(_count_tests, string="Tests no OK",
+                                        type="integer", multi="test_count"),
+    }
+
+    # @api.cr_uid_ids_context  # this is taken from stock module
+    def do_transfer(self, cr, uid, ids, context=None):
+        do_transfer = super(StockPicking, self).do_transfer(cr, uid, ids,
+                                                            context=context)
+        if do_transfer:
+            for picking in self.browse(cr, uid, ids, context=context):
+                if (picking.state == 'done' and
+                        picking.picking_type_id.code == 'incoming' and
+                        picking.move_lines):
+                    move_obj = self.pool['stock.move']
+                    for move in picking.move_lines:
+                        how_many = move_obj._how_many_test_create(
+                            cr, uid, move.product_id, move.product_qty,
+                            context=context)
+                        if how_many > 0:
+                            move_obj._create_test_automatically(
+                                cr, uid, how_many, move, context=context)
+        return do_transfer
+
+
+class StockProductionLot(orm.Model):
+    _inherit = 'stock.production.lot'
+
+    def _count_tests(self, cr, uid, ids, name, arg, context=None):
+        res = {}
+        for test in self.browse(cr, uid, ids, context=context):
+            created = 0
+            done = 0
+            passed = 0
+            failed = 0
+            if test.qc_test_ids:
+                for line in test.qc_test_ids:
+                    created += 1
+                    if line.state not in ('draft', 'waiting'):
+                        done += 1
+                        if line.state == 'success':
+                            passed += 1
+                        elif line.state == 'failed':
+                            failed += 1
+            res[test.id] = {
+                'created_tests': created,
+                'done_tests': done,
+                'passed_tests': passed,
+                'failed_tests': failed,
+            }
+        return res
+
+    _columns = {
+        # Test asociados al albaran
+        'qc_test_ids': fields.one2many('qc.test', 'lot_id', 'Tests'),
+        # Numero de test creados
+        'created_tests': fields.function(_count_tests, string="Created Tests",
+                                         type="integer", multi="test_count"),
+        # Numero de test realizados
+        'done_tests': fields.function(_count_tests, string="Realized Tests",
+                                      type="integer", multi="test_count"),
+        # Numero de test OK
+        'passed_tests': fields.function(_count_tests, string="Tests OK",
+                                        type="integer", multi="test_count"),
+        # Numero de test fallados
+        'failed_tests': fields.function(_count_tests, string="Tests no OK",
+                                        type="integer", multi="test_count"),
+    }
