--- quality_control_ext/models/stock_picking.py
@@ -1,86 +1,0 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    Avanzosc - Advanced Open Source Consulting
-#    Copyright (C) 2011 - 2013 Avanzosc <http://www.avanzosc.com>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see http://www.gnu.org/licenses/.
-#
-##############################################################################
-
-from openerp.osv import orm, fields
-
-
-class StockPicking(orm.Model):
-    _inherit = 'stock.picking'
-
-    def _count_tests(self, cr, uid, ids, name, arg, context=None):
-        res = {}
-        for test in self.browse(cr, uid, ids, context=context):
-            created = 0
-            done = 0
-            passed = 0
-            failed = 0
-            if test.qc_test_ids:
-                for line in test.qc_test_ids:
-                    created += 1
-                    if line.state not in ('draft', 'waiting'):
-                        done += 1
-                        if line.state == 'success':
-                            passed += 1
-                        elif line.state == 'failed':
-                            failed += 1
-            res[test.id] = {
-                'created_tests': created,
-                'done_tests': done,
-                'passed_tests': passed,
-                'failed_tests': failed,
-            }
-        return res
-
-    _columns = {
-        # Test asociados al albaran
-        'qc_test_ids': fields.one2many('qc.test', 'picking_id', 'Tests'),
-        # Numero de test creados
-        'created_tests': fields.function(_count_tests, string="Created Tests",
-                                         type="integer", multi="test_count"),
-        # Numero de test realizados
-        'done_tests': fields.function(_count_tests, string="Realized Tests",
-                                      type="integer", multi="test_count"),
-        # Numero de test OK
-        'passed_tests': fields.function(_count_tests, string="Tests OK",
-                                        type="integer", multi="test_count"),
-        # Numero de test fallados
-        'failed_tests': fields.function(_count_tests, string="Tests no OK",
-                                        type="integer", multi="test_count"),
-    }
-
-    # @api.cr_uid_ids_context  # this is taken from stock module
-    def do_transfer(self, cr, uid, ids, context=None):
-        do_transfer = super(StockPicking, self).do_transfer(cr, uid, ids,
-                                                            context=context)
-        if do_transfer:
-            for picking in self.browse(cr, uid, ids, context=context):
-                if (picking.state == 'done' and
-                        picking.picking_type_id.code == 'incoming' and
-                        picking.move_lines):
-                    move_obj = self.pool['stock.move']
-                    for move in picking.move_lines:
-                        how_many = move_obj._how_many_test_create(
-                            cr, uid, move.product_id, move.product_qty,
-                            context=context)
-                        if how_many > 0:
-                            move_obj._create_test_automatically(
-                                cr, uid, how_many, move, context=context)
-        return do_transfer
